removed deprecated code
